added geocoding
added geocoding

+++ db/migrate/20151013110109_create_reviews.rb
@@ -2,8 +2,8 @@ class CreateReviews < ActiveRecord::Migration
   def change
     create_table :reviews do |t|
       t.integer :user_id
-      t.decimal :lat, :precision=>64, :scale=>12
-      t.decimal :lon, :precision=>64, :scale=>12
+      t.double :lat, :precision=>64, :scale=>12
+      t.double :lon, :precision=>64, :scale=>12
       t.string :title
       t.text :description
       t.string :question1
